Change namespace.
