+++ src/Dali/Dali.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29609.76
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali", "Dali\RedSharp.Dali.csproj", "{AC71636F-8BA8-4D51-8865-F15950D13995}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali.View", "Dali\RedSharp.Dali.View.csproj", "{AC71636F-8BA8-4D51-8865-F15950D13995}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedSharp.Dali.ViewModel", "RedSharp.Dali.ViewModel\RedSharp.Dali.ViewModel.csproj", "{2F5CA780-944A-4918-A8FD-A4CE6DE99D98}"
 EndProject
