+++ src/Dali/Dali/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application x:Class="RedSharp.Dali.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:converters="clr-namespace:RedSharp.Dali.Converters">
+             xmlns:converters="clr-namespace:RedSharp.Dali.View.Converters">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
