+++ src/Dali/Dali/Resources/Strings/Strings.xaml
@@ -1,6 +1,5 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:RedSharp.Dali.Resources.Strings"
                     xmlns:sys="clr-namespace:System;assembly=mscorlib">
 
     <sys:String x:Key="WindowTitle">Dali</sys:String>
