--- src/Dali/Dali/View/MainWindow.xaml.cs
@@ -1,34 +1,0 @@
-﻿using RedSharp.Dali.Controls.Windows;
-using RedSharp.Dali.ViewModel;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using Unity;
-
-namespace RedSharp.Dali.View
-{
-    public partial class MainWindow : DaliWindow
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-
-        [Dependency]
-        public MainWindowViewModel ViewModel
-        {
-            set { DataContext = value; }
-        }
-    }
-}
