+++ src/Dali/Dali/Windows/MainWindow.xaml
@@ -1,12 +1,12 @@
-﻿<controls:DaliWindow x:Class="RedSharp.Dali.View.MainWindow"
+﻿<controls:DaliWindow x:Class="RedSharp.Dali.View.Windows.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:controls="clr-namespace:RedSharp.Dali.Controls.Windows;assembly=RedSharp.Dali.Controls"
         xmlns:viewModels="clr-namespace:RedSharp.Dali.ViewModel;assembly=RedSharp.Dali.ViewModel"
-        xmlns:actions="clr-namespace:RedSharp.Dali.Actions"
-        xmlns:behaviors="clr-namespace:RedSharp.Dali.Behaviors"
+        xmlns:actions="clr-namespace:RedSharp.Dali.View.Actions"
+        xmlns:behaviors="clr-namespace:RedSharp.Dali.View.Behaviors"
         xmlns:i="http://schemas.microsoft.com/xaml/behaviors" 
         FocusManager.FocusedElement="{Binding ElementName=LoadButton}"
         mc:Ignorable="d" Icon="{StaticResource AppIcon}"
