+++ src/Dali/Dali/Windows/MainWindow.xaml.cs
@@ -1,0 +1,20 @@
+﻿using RedSharp.Dali.Controls.Windows;
+using RedSharp.Dali.ViewModel;
+using Unity;
+
+namespace RedSharp.Dali.View.Windows
+{
+    public partial class MainWindow : DaliWindow
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+        }
+
+        [Dependency]
+        public MainWindowViewModel ViewModel
+        {
+            set { DataContext = value; }
+        }
+    }
+}
