+++ src/Dali/Dali/Windows/TransparentWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="RedSharp.Dali.View.TransparentWindow"
+﻿<Window x:Class="RedSharp.Dali.View.Windows.TransparentWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
