+++ src/Dali/Dali/Windows/TransparentWindow.xaml.cs
@@ -1,19 +1,8 @@
 ﻿using RedSharp.Dali.Common.Interop.Helpers;
-using RedSharp.Dali.Controls.Windows;
 using RedSharp.Dali.ViewModel;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
-namespace RedSharp.Dali.View
+namespace RedSharp.Dali.View.Windows
 {
     public partial class TransparentWindow : Window
     {
